% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_sf.R
\name{df_geojson}
\alias{df_geojson}
\title{df to GeoJSON}
\usage{
df_geojson(df, lon, lat, atomise = FALSE, simplify = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{atomise}{logical indicating if the data.frame should be converted into a vector
of GeoJSON objects}

\item{simplify}{logical indicating if data.frame without property columns should simplify
(\code{TRUE}) into a vector of GeoJSON, or (\code{FALSE}). If \code{atomise} is TRUE
this argument is ignored.}
}
\value{
vector of GeoJSON
}
\description{
Converts data.frame objects to GeoJSON
}
\examples{
\dontrun{

df <- data.frame(lon = 1:5, lat = 1:5, id = 1:5, val = letters[1:5])
df_geojson( df, lon = "lon", lat = "lat")
df_geojson( df, lon = "lon", lat = "lat", atomise = T)

df <- data.frame(lon = 1:5, lat = 1:5)
df_geojson( df, lon = "lon", lat = "lat")
df_geojson( df, lon = "lon", lat = "lat", simplify = T)


}

}
