% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonsf-package.R
\docType{data}
\name{geo_melbourne}
\alias{geo_melbourne}
\title{geo_melbourne}
\format{An object of class \code{json} of length 1.}
\usage{
geo_melbourne
}
\description{
GeoJSON data of Melbourne's Inner suburbs.
}
\keyword{datasets}
