% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_sf.R
\name{geojson_sf}
\alias{geojson_sf}
\title{Geojson to sf}
\usage{
geojson_sf(geojson)
}
\arguments{
\item{geojson}{string or vector of GeoJSON}
}
\description{
Converts GeoJSON to an `sf` object
}
\examples{

## load 'sf' for print methods
# library(sf)
geojson <- '{ "type" : "Point", "coordinates" : [0, 0] }'
geojson_sf(geojson)

}
