% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsf_sf.R
\name{geojson_sfc}
\alias{geojson_sfc}
\title{Geojson to sfc}
\usage{
geojson_sfc(geojson)
}
\arguments{
\item{geojson}{string of GeoJSON}
}
\description{
Extracts geometries from GeoJSON and returns an `sfc` object
}
