% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_wkt.R
\name{geojson_wkt}
\alias{geojson_wkt}
\title{Geojson to WKT}
\usage{
geojson_wkt(geojson)
}
\arguments{
\item{geojson}{string or vector of GeoJSON, or a URL or file pointing to a geojson file}
}
\value{
data.frame with a 'geometry' column of well-known text
}
\description{
Converts GeoJSON to Well-Known Text
}
\examples{
geojson <- '{ "type" : "Point", "coordinates" : [0, 0] }'
geojson_wkt(geojson)

}
