% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_sf.R
\name{df_geojson}
\alias{df_geojson}
\title{df to GeoJSON}
\usage{
df_geojson(df, lon, lat, atomise = FALSE, simplify = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{atomise}{logical indicating if the data.frame should be converted into a vector
of GeoJSON objects}

\item{simplify}{logical indicating if data.frame without property columns should simplify
(\code{TRUE}) into a vector of GeoJSON, or (\code{FALSE}). If \code{atomise} is TRUE
this argument is ignored.}
}
\value{
vector of GeoJSON
}
\description{
Converts data.frame objects to GeoJSON
}
\examples{
\dontrun{
library(sf)
sf <- sf::st_sf(geometry = sf::st_sfc(list(sf::st_point(c(0,0)), sf::st_point(c(1,1)))))
sf$id <- 1:2
sf_geojson(sf)
sf_geojson(sf, atomise = T)

ls <- st_linestring(rbind(c(0,0),c(1,1),c(2,1)))
mls <- st_multilinestring(list(rbind(c(2,2),c(1,3)), rbind(c(0,0),c(1,1),c(2,1))))
sfc <- st_sfc(ls,mls)
sf <- st_sf(sfc)
sf_geojson( sf )
sf_geojson( sf, simplify = FALSE )

}

}
