% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_sf.R
\name{geojson_sf}
\alias{geojson_sf}
\title{Geojson to sf}
\usage{
geojson_sf(geojson)
}
\arguments{
\item{geojson}{string or vector of GeoJSON, or a URL or file pointing to a geojson file}
}
\description{
Converts GeoJSON to an `sf` object
}
\examples{

## character string of GeoJSON

## load 'sf' for print methods
# library(sf)
geojson <- '{ "type" : "Point", "coordinates" : [0, 0] }'
geojson_sf(geojson)


\dontrun{
## GeoJSON at a url
myurl <- "http://eric.clst.org/assets/wiki/uploads/Stuff/gz_2010_us_050_00_500k.json"
sf <- geojson_sf(myurl)
}

}
