% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_sf.R
\name{geojson_sfc}
\alias{geojson_sfc}
\title{Geojson to sfc}
\usage{
geojson_sfc(geojson)
}
\arguments{
\item{geojson}{string or vector of GeoJSON, or a URL or file pointing to a geojson file}
}
\description{
Extracts geometries from GeoJSON and returns an `sfc` object
}
\examples{
## load 'sf' for print methods
# library(sf)
geojson <- '{ "type" : "Point", "coordinates" : [0, 0] }'
geojson_sfc(geojson)

}
