% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_sf.R
\name{sf_geojson}
\alias{sf_geojson}
\title{sf to GeoJSON}
\usage{
sf_geojson(sf, atomise = FALSE)
}
\arguments{
\item{sf}{simple feature object}

\item{atomise}{logical}
}
\description{
Converts `sf`, `sfc` and `sfg` objects to GeoJSON
}
