% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_sf.R
\name{sf_geojson}
\alias{sf_geojson}
\title{sf to GeoJSON}
\usage{
sf_geojson(sf, atomise = FALSE)
}
\arguments{
\item{sf}{simple feature object}

\item{atomise}{logical}
}
\value{
vector of GeoJSON
}
\description{
Converts `sf` objects to GeoJSON
}
\examples{
\dontrun{
library(sf)
sf <- sf::st_sf(geometry = sf::st_sfc(list(sf::st_point(c(0,0)), sf::st_point(c(1,1)))))
sf$id <- 1:2
sf_geojson(sf)
sf_geojson(sf, atomise = T)
}

}
