% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_sf.R
\name{sfc_geojson}
\alias{sfc_geojson}
\title{sfc to GeoJSON}
\usage{
sfc_geojson(sfc)
}
\arguments{
\item{sfc}{simple feature collection object}
}
\value{
vector of GeoJSON
}
\description{
Converts `sfc` objects to GeoJSON
}
\examples{
\dontrun{
library(sf)
sf <- sf::st_sfc(list(sf::st_point(c(0,0)), sf::st_point(c(1,1))))
sfc_geojson(sf)
}
}
