% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_geojson.R
\name{sfc_geojson}
\alias{sfc_geojson}
\title{sfc to GeoJSON}
\usage{
sfc_geojson(sfc, digits = NULL)
}
\arguments{
\item{sfc}{simple feature collection object}

\item{digits}{integer specifying the number of decimal places to round numeric coordinates.
numeric values are coorced using \code{as.integer}, which may round-down the value you supply.
Default is \code{NULL} - no rounding}
}
\value{
vector of GeoJSON
}
\description{
Converts `sfc` objects to GeoJSON
}
\examples{
\dontrun{
library(sf)
sf <- sf::st_sfc(list(sf::st_point(c(0,0)), sf::st_point(c(1,1))))
sfc_geojson(sf)
}
}
